matching components CalendarChief and ShiftForm and create one ShiftScheduler instead of those two. Update App.jsx
